refactor: rename to useErrorHandler.ts
